#121 Add main.py
