+++ data/ckan/Perth and Kinross Council.csv
@@ -102,8 +102,8 @@ Employment Land Audit 2022,Perth & Kinross Council,https://data.pkc.gov.uk/datas
 Employment Land Audit 2022,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/employment-land-audit-2022,https://open-data-perth-kinross.hub.arcgis.com/datasets/576594772a76474980e4ce11cc2ea5a3_9.geojson?outSR=%7B%22latestWkid%22%3A27700%2C%22wkid%22%3A27700%7D,GeoJSON,2023-03-29T14:41:29.991750,2023-03-29T14:42:40.678769,0,B,GeoJSON,,2022;ELA;Economy and Business,,,<div style='text-align:Left;font-size:12pt'><div><div><p><span>The Perth and Kinross Employment Land Audit</span><span> </span><span>has been prepared from information provided in the Local Development Plan and through the monitoring of planning application approvals.</span></p></div></div></div>
 Employment Land Audit 2022,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/employment-land-audit-2022,https://open-data-perth-kinross.hub.arcgis.com/datasets/576594772a76474980e4ce11cc2ea5a3_9.zip?outSR=%7B%22latestWkid%22%3A27700%2C%22wkid%22%3A27700%7D,Shapefile,2023-03-29T14:41:29.991750,2023-03-29T14:42:40.678769,0,B,ZIP,,2022;ELA;Economy and Business,,,<div style='text-align:Left;font-size:12pt'><div><div><p><span>The Perth and Kinross Employment Land Audit</span><span> </span><span>has been prepared from information provided in the Local Development Plan and through the monitoring of planning application approvals.</span></p></div></div></div>
 Employment Land Audit 2022,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/employment-land-audit-2022,https://open-data-perth-kinross.hub.arcgis.com/datasets/576594772a76474980e4ce11cc2ea5a3_9.kml?outSR=%7B%22latestWkid%22%3A27700%2C%22wkid%22%3A27700%7D,KML,2023-03-29T14:41:29.991750,2023-03-29T14:42:40.678769,0,B,KML,,2022;ELA;Economy and Business,,,<div style='text-align:Left;font-size:12pt'><div><div><p><span>The Perth and Kinross Employment Land Audit</span><span> </span><span>has been prepared from information provided in the Local Development Plan and through the monitoring of planning application approvals.</span></p></div></div></div>
-Energy - Electricity,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/energy-electricity,https://data.pkc.gov.uk/dataset/3d8d3d43-35fc-4da1-b52b-0daf6c9c6f72/resource/ed1fd7df-69c5-41a3-b52f-ca111a308d34/download/energyelectricity.csv,EnergyElectricityPKC,2019-09-10T11:04:08.821142,2023-03-26T13:01:07.176527,16775521,B,CSV,,Energy,,Open Government Licence 3.0 (United Kingdom),Electricity usage data from Perth & Kinross Council owned buildings that have meters where automatic readings are recorded.  Data is for the last  2 years.
-Energy - Gas,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/energy-gas,https://data.pkc.gov.uk/dataset/fc4db049-1f17-4e35-82fc-8ccf8d101ec1/resource/9b1271cf-5b26-4b29-9468-efb22cbf1256/download/energygas.csv,EnergyGasPKC,2019-09-10T12:11:10.539934,2023-03-26T13:01:31.518721,23138889,B,CSV,,energy,,Open Government Licence 3.0 (United Kingdom),Gas usage data from Perth & Kinross Council owned buildings fitted with meters that record automatic readings. Data is for the last 2 years.
+Energy - Electricity,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/energy-electricity,https://data.pkc.gov.uk/dataset/3d8d3d43-35fc-4da1-b52b-0daf6c9c6f72/resource/ed1fd7df-69c5-41a3-b52f-ca111a308d34/download/energyelectricity.csv,EnergyElectricityPKC,2019-09-10T11:04:08.821142,2023-04-02T13:01:16.353249,16747603,B,CSV,,Energy,,Open Government Licence 3.0 (United Kingdom),Electricity usage data from Perth & Kinross Council owned buildings that have meters where automatic readings are recorded.  Data is for the last  2 years.
+Energy - Gas,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/energy-gas,https://data.pkc.gov.uk/dataset/fc4db049-1f17-4e35-82fc-8ccf8d101ec1/resource/9b1271cf-5b26-4b29-9468-efb22cbf1256/download/energygas.csv,EnergyGasPKC,2019-09-10T12:11:10.539934,2023-04-02T13:01:41.838442,23175373,B,CSV,,energy,,Open Government Licence 3.0 (United Kingdom),Gas usage data from Perth & Kinross Council owned buildings fitted with meters that record automatic readings. Data is for the last 2 years.
 Electric Vehicle Charging Station Usage,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/ev-charging-data,https://data.pkc.gov.uk/dataset/a3402e38-e9a5-4917-a23b-cbd18971754d/resource/d0bc0302-90e5-4f8c-bc2e-b54e648b88b6/download/electricvehiclechargecorrected.csv,September 2016 - August 2017,2017-09-15T13:03:10.781384,2021-12-15T09:02:47.974017,1370792,B,CSV,,energy,,Open Government Licence 3.0 (United Kingdom),Datasets for Perth & Kinross Council's EV charging stations under the ChargePlace Scotland scheme. Includes anonymous data from each individual charging session.
 Electric Vehicle Charging Station Usage,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/ev-charging-data,https://data.pkc.gov.uk/dataset/a3402e38-e9a5-4917-a23b-cbd18971754d/resource/f5f95c50-2ca1-436d-bd6e-1c5f8317ac9b/download/sept17toaug18standardisedcorrected.csv,September 2017-August 2018,2017-09-15T13:03:10.781384,2021-12-15T09:02:47.974017,2140941,B,CSV,,energy,,Open Government Licence 3.0 (United Kingdom),Datasets for Perth & Kinross Council's EV charging stations under the ChargePlace Scotland scheme. Includes anonymous data from each individual charging session.
 Electric Vehicle Charging Station Usage,Perth & Kinross Council,https://data.pkc.gov.uk/dataset/ev-charging-data,https://data.pkc.gov.uk/dataset/a3402e38-e9a5-4917-a23b-cbd18971754d/resource/845fe22f-19f4-499b-b8bc-8812a763e660/download/sept18toaug19standardisedcorrected.csv,September 2018-August 2019,2017-09-15T13:03:10.781384,2021-12-15T09:02:47.974017,3417102,B,CSV,,energy,,Open Government Licence 3.0 (United Kingdom),Datasets for Perth & Kinross Council's EV charging stations under the ChargePlace Scotland scheme. Includes anonymous data from each individual charging session.
