+++ log.md
@@ -1,0 +1,7 @@
+# pipeline error log
+
+## Unaccessible Webpages
+
+|URL | Error Code | Error Reason|
+|--- | --- | ---|
+| https://data.stirling.gov.uk//api/3/action/package_list | 404 | Not Found |
